graph, graphql: Handle selection/leaf mismatch
Make sure that we handle queries that have a selection from a scalar
field by ignoring the selection or that have no selection for a non-scalar
field by ignoring that field.

The latter differs from the previous behavior where the result would
contain an entry for such a field, but the data for the field would be an
empty object or a list of empty objects.